Bug fixes in: 1. Computing max ghost cells for GPU, 2. Invalidating staging variables after "modifies". Also promoted KScalarRHS to GPU.
1. Computing max ghost cells for GPU: Aded a flag in SchedulerCommon to
indicate max ghost cell compute phase. If the flag is set to true, then
tasks are just parsed, but not added into the task graph.
2. Invalidating staging variables after "modifies": Temporary fix -
Commented "((oldVarStatus & VALID) == VALID)" condition in
GPUDataWarehouse::compareAndSwapCopyingIntoGPUStaging as a temporary fix
for the defect. When a variable is modified on GPU, ValidWithGhost
status is reverted to allow gathering of ghost cells again for the next
requires dependency. But as of now there is no mechanism to mark staging
variables invalid. As a result, although the ghost cells on the main
variable are invalidated, staging variables still have valid status and
hold old values and because of valid status prepareDeviceVars does not
issue fresh H2D copy for the staging variable. The permanent fix is to
find out and inactivate staging variables of neighboring patches. But
that needs more time. So this temporary fix to ignore valid status as of
now which will cause few redundant h2d copies, but will make code work.