* Updated default scheduler type logic. * Updated scheduler type bulletproofing.
This commit allows scheduler type "Kokkos" to be set based on run-time
parameters passed when executing sus.

Scheduler type "Kokkos" maps to the Kokkos Scheduler implemented in
KokkosScheduler.cc/h.

This scheduler is based on the Unified Scheduler and aims to support
simultaneous use of the Kokkos::OpenMP and Kokkos::CUDA back-ends.

The fundamental difference between the Kokkos Scheduler and the Unified
Scheduler is that the Kokkos Scheduler replaces the Unified Scheduler's
std::thread-based TaskRunners with OpenMP-based equivalents.

This scheduler can be invoked explicitly from within an input using the
XML block below:

<Scheduler type="Kokkos">
</Scheduler>

This scheduler can be invoked implicitly at run-time using the sus
options below:

-npartitions <#> -nthreadsperpartition <#> -gpu