Approaches to run RMCRT and Hypre together with optimal patch size
RMCRT runs more efficiently with smaller patches because patch based ROI
decreases with patch size and rays jump to coarser levels sooner. Thus
RMCRT needs division of the sub-domain assigned to the rank into smaller
patches. Hypre-cuda works the best with patche size >= 128 cubed. Thus
the two tasks conflict with each other.

Two experimental approaches are used here to solve the problem. Both can
be used individually or together:
1. Superpatch for hypre: Hypre task takes Low Index of the first patch
and High Index of the last patch assigned to the rank and passes a giant
super patch to Hypre. Assumption is the rank's subdomain is contiguous
and aligned with the domain boundaries. It allows RMCRT to run with
smaller patches and combines patches within Hypre to achieve larger
patch size. Set environment variable HYPRE_SUPERPATCH to 1 to activate.
2. "virtual patch size" for RMCRT's ROI: Irrespective of  the real patch
size, ROI will be computed assuming the patch size given by
VIR_ROI_SIZE. e.g. with VIR_ROI_SIZE=16,16,16 RMCRT task will compute
ROI boundaries assuming patch size is 16 cubed. This allows setting
actual patches to be really big, but smaller ROI boundaries give the
same efficiency (and the accuracy) as if the patches are small. Set the
environment variable VIR_ROI to 1 and VIR_ROI_SIZE to a comma seperated
list of expected patch size (not the number of patches) for ROI.

Can be written using ups tags later on if these work well.