Two big changes.  Adding a "clay" attribute to the triangles as the presence of a clay coating affects dissolution rate.  So, models were changed to accomodate that attribute as well.
Also, adding, optionally for now, the ability to use a time averaged kinetic energy in comparing against KE thresholds for determining when to increase load,
or modify load curve.  Also, this is a "specific" KE, that is, KE normalized
by piston mass.  This will allow use of the same KE threshold regardless of
piston size.