Another fix for MPI conflicts between Hypre and Uintah
On few occasions while running on Frontera, one rank started Hypre task
and was watining on MPI_reduce within Hypre, while other ranks started
Uintah's reduction tasks waiting for MPI_reduce in it. So added
task->usesMPI(true) in hypre task with a hope that the taskgraph and the
scheduler will create a different phase for Hypre task and avoid these
conflicts. Fixed scheduler logic to backtrack for hypre task to include
phases.