Simplified the code for the compression of variables before writing to the uda. Previously, if user specified:
 <compression> gzip </compression>

there was logic (circa 2005) to determine to write either compressed or uncompressed variables to the uda.
This occured for every variable, patch and timestep.  An exception was thrown when reading an uda
if the variable was supposed to be compressed but it wasn't or vice versa.

The logic has been removed and every variable on every patch and timestep is compressed.